+++ splineABS/R/gen_subset.R
@@ -1,24 +1,48 @@
 #' @title Generate sampling basis in each subset
 #'
-#' @description  The subsets is divied randomly and evenly. Return a list of indices of samples and indices basis in each subset.
+#' @description  The subsets is divied randomly and evenly.
+#'   Return a list of indices of samples and indices basis in each subset.
 #'
 #' @param N             The number of whole smaples.
-#' @param sample.index  The indices of samples.
+#' @param basis.index  The indices of samples. It could be NULL.
 #' @param subset.number The number of subsets.
 #' @param ratio         A vector of the ratio of size of each subset. Default is equal.
 #'
-#' @return       A list contains two vectors. The indices of samples and basis in each subset.
+#' @return  A list contains two lists named as \code{subsetElementIndex}, \code{basisIndexInSubset}.
+#' \itemize{
+#'   \item \code{subsetElementIndex}: The global indices(global location) of each subset's elements in whole data.
+#'   \item \code{basisIndexInSubset} The local indices(local location) of basis in each subset.
+#'   \item remark: global means the location in whole data; local means the location in each subset.
+#' }
+#'
+#'
+#'
 #'
 #' @examples
-#' generate.subset(y = 1:10, sample.index = c(1,3,5), subset.number = 2)
+#' > generate.subset(20, basis.index = c(1,6, 11,16), subset.number = 2)
+#' $subsetElementIndex
+#' $subsetElementIndex[[1]]
+#' [1]  1  2  3  5  6  7  9 11 12 13 14 16 18
+#'
+#' $subsetElementIndex[[2]]
+#' [1]  1  4  6  8 10 11 15 16 17 19 20
+#'
+#'
+#' $basisIndexInSubset
+#' $basisIndexInSubset[[1]]
+#' [1]  1  5  8 12
+#'
+#' $basisIndexInSubset[[2]]
+#' [1] 1 3 6 8
 #'
 #' @export
 #'
-generate.subset <- function(N, sample.index, subset.number=1, ratio = NULL){
+generate.subset <- function(N,basis.index = NULL, subset.number = 1, ratio = NULL){
 
   whole.index <- 1:N
-  subset.index <- list()
-  subset.sample.index <- list()
+  subset.element.index <- list()
+  basis.index.in.subset <- list()
+
   if(is.null(ratio)){
     # get a rough subset size
     subset.meansize <- round(N / subset.number)
@@ -32,7 +56,6 @@ generate.subset <- function(N, sample.index, subset.number=1, ratio = NULL){
     # get an exact subset size (adjust the largest subset)
     subset.size[index.largest.subset] <- subset.size[index.largest.subset] -
       (sum(subset.size) - N)
-    print(subset.size)
   }
 
 
@@ -40,14 +63,25 @@ generate.subset <- function(N, sample.index, subset.number=1, ratio = NULL){
 
   for(i in 1:(subset.number)){
 
+    # 1) generate subsets
     if(i != subset.number){
-      subset.index[[i]] <- sort(sample(remain.index, subset.size[i]))
-      remain.index <- setdiff(remain.index, subset.index[[i]])
+      subset.element.index[[i]] <- sample(remain.index, subset.size[i])
+      remain.index <- setdiff(remain.index, subset.element.index[[i]])
     }else{
-      subset.index[[i]] <- sort(remain.index)
+      subset.element.index[[i]] <- remain.index
     }
-    subset.sample.index[[i]] <- intersect(sample.index, subset.index[[i]])
+
+    # 2) copy basis into each subset
+    # ganrantee that each subset shares the same basis
+    subset.element.index[[i]] <- sort(union(subset.element.index[[i]], basis.index))
+
+    # 3) record the location of basis in each subset
+
+      basis.index.in.subset[[i]] <- match(basis.index, subset.element.index[[i]])
   }
 
-  return(list(subset.index=subset.index, subset.sample.index=subset.sample.index))
+  # return a list: 1) indices(global location) of subset elements;
+  # 2) local indices(local location) of basis in each subset
+  return(list(subsetElementIndex = subset.element.index, basisIndexInSubset = basis.index.in.subset))
+
 }